git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fault-localization-import@42416 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/pathformula/pointeraliasing/PointerTargetSetManager.java
@@ -70,7 +70,7 @@ class PointerTargetSetManager {
    * @param size The size of the fake base type.
    * @return An array of {@code size} voids.
    */
-  static CType getFakeBaseType(int size) {
+  static CType getFakeBaseType(long size) {
     return checkIsSimplified(
         new CArrayType(
             false,
@@ -568,7 +568,7 @@ class PointerTargetSetManager {
               false));
     }
 
-    final int typeSize =
+    final long typeSize =
         pType.hasKnownConstantSize()
             ? typeHandler.getSizeof(pType)
             : options.defaultAllocationSize();
@@ -672,7 +672,7 @@ class PointerTargetSetManager {
     if (cType instanceof CArrayType) {
       final CArrayType arrayType = (CArrayType) cType;
       final int length = CTypeUtils.getArrayLength(arrayType, options);
-      int offset = 0;
+      long offset = 0;
       for (int i = 0; i < length; ++i) {
         // TODO: create region with arrayType.getType()
         targets =
